Add files via upload
As we can see in the report that the insight has been made on the basis of previous five years data including values ( in rupees) being spent on defence sector by Indian government all these years, collected and analysed to calculate future value (in rupees) to be spent in future i.e. 2026-27.